+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## HEAD
+- Implemented Refs, Events and Graph directly within the KeyValue class, check out our README.
+- next/prev methods are now named as nextPage/prevPage, much clearer on what they do.
+
 ## 0.8.0 - 2015-03-06
 - Radical change on classes, Application is now rightfully 'Client'.
 - Application is defined as the entry point to the object API.